requirements.txt
# local package
-e .

# external requirements
click~=8.0.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.4.0
requests~=2.27.1
setuptools~=57.0.0
lxml~=4.7.1
beautifulsoup4~=4.10.0
tabulate~=0.8.9
numpy~=1.22.1